Refactored folder structure & checked to see homepage, login, sign up, and locations. Continued work needed to render profile pages
