+++ controllers/api/index.js
@@ -1,8 +1,8 @@
 const router = require("express").Router();
 const userRoutes = require("./userRoutes");
-const projectRoutes = require("./locationsRoutes");
+const locationRoutes = require("./locationsRoutes");
 
 router.use("/users", userRoutes);
-router.use("/projects", projectRoutes);
+router.use("/locations", locationRoutes);
 
 module.exports = router;
