+++ controllers/homeRoutes.js
@@ -1,9 +1,9 @@
 const router = require("express").Router();
-const { User } = require("../../models");
-const withAuth = require("../../utils/auth");
+const { User } = require("../models");
+const withAuth = require("../utils/auth");
 
 // Prevent non logged in users from viewing the homepage
-router.get("/home", withAuth, async (req, res) => {
+router.get("/", withAuth, async (req, res) => {
   try {
     const userData = await User.findAll({
       attributes: { exclude: ["password"] },
@@ -16,6 +16,25 @@ router.get("/home", withAuth, async (req, res) => {
       users,
       // Pass the logged in flag to the template
       loggedIn: req.session.loggedIn,
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+router.get('/profile', withAuth, async (req, res) => {
+  try {
+    // Find the logged in user based on the session ID
+    const userData = await User.findByPk(req.session.id, {
+      attributes: { exclude: ['password'] },
+      include: [{ model: User }],
+    });
+
+    const user = userData.get({ plain: true });
+
+    res.render('profile', {
+      ...user,
+      logged_in: true
     });
   } catch (err) {
     res.status(500).json(err);
