+++ views/homepage.handlebars
@@ -1,1 +1,4 @@
 {{!-- Renders globe explorer thing --}}
+<div>
+    It works!
+</div>
