+++ views/loginpage.handlebars
@@ -13,3 +13,22 @@
     <button class="btn btn-primary" type="submit">login</button>
   </div>
 </form>
+
+<h2>Sign-Up</h2>
+<form class="form signup-form">
+   <div class="form-group">
+    <label for="name-signup">name:</label>
+    <input class="form-input" type="text" id="name-signup"/>
+  </div>
+  <div class="form-group">
+    <label for="email-singup">email:</label>
+    <input class="form-input" type="text" id="email-signup"/>
+  </div>
+  <div class="form-group">
+    <label for="password-signup">password:</label>
+    <input class="form-input" type="password" id="password-signup" />
+  </div>
+  <div class="form-group">
+    <button class="btn btn-primary" type="submit">sign up</button>
+  </div>
+</form>
