+++ views/profile.handlebars
@@ -1,0 +1,3 @@
+<div>
+    It works, welcome to your profile page!
+</div>
